ai/models/Stable-Video-Diffusion/requirements.txt
fastapi
uvicorn[standard]
torch
diffusers
transformers
accelerate
Pillow
